requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
requests
pillow
langchain
yandex-chain
opensearch-py
boto3
pandas
pydantic
langchain-community
loguru>=0.7.0
grpcio>=1.59.0
yandexcloud>=0.254.0